improve StreetView pano guess position
previous approach getPanoramaByLocation returned many cases with
indoor imaging from a neighboring store. Instead, now we use jquery to
call a python function which uses an unofficial Google API that
returns the best-guess outdoor pano ID. This should be used as a
starting point for StreetsAhead location/orientation as well.