add appendCSVToTable to add old data from yelp db to main db
coffeeshops.py script places all cached calls and image values into
yelp database with same table format as main StreetsAhead db. Tried to
export yelp db dumb as csv and add to main SQL db, but most of the
calls used 5 pointings instead of 3 so headings are different and the
value of caching isn't that great.